# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your super  ApplicationAPI",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000"
        }
    ],
    "paths": {
        "/employee/attendance/{employee_id}": {
            "get": {
                "tags": [
                    "Attendance"
                ],
                "summary": "Get attendance information for an employee",
                "description": "Get attendance information for an employee including total working hours.",
                "operationId": "getAttendanceInformation",
                "parameters": [
                    {
                        "name": "employee_id",
                        "in": "path",
                        "description": "Employee ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "Start date (YYYY-MM-DD)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "End date (YYYY-MM-DD)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Employee not found"
                    }
                }
            }
        },
        "/attendance": {
            "get": {
                "tags": [
                    "Attendance"
                ],
                "summary": "Get a list of all attendance records",
                "operationId": "index",
                "responses": {
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/attendance": {
            "post": {
                "tags": [
                    "Attendance"
                ],
                "summary": "Import attendance data from an Excel file",
                "operationId": "store",
                "requestBody": {
                    "description": "Excel file to import",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "file": {
                                        "description": "The Excel file to import",
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Attendance data imported successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/group-by-owners": {
            "get": {
                "tags": [
                    "group-by-owners"
                ],
                "summary": "Group files by owners",
                "operationId": "988e345f118b40e1c00ff63eaf46d23c",
                "requestBody": {
                    "description": "Input data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "insurance.txt": {
                                        "type": "string",
                                        "example": "Company A"
                                    },
                                    "letter.docx": {
                                        "type": "string",
                                        "example": "Company A"
                                    },
                                    "Contract.docx": {
                                        "type": "string",
                                        "example": "Company B"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": " successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        }
    }
}